Modified Astrolabe CLI to move functionality into pkg/cmd/cli_core.go and allow for adding additional plugins (can be used from a external package that include non-core PEs)
